requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
frozenlist==1.3.3
gcloud==0.18.3
googleapis-common-protos==1.58.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
oauth2client==4.1.3
protobuf==4.22.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.13.0
